Split guest auth test harness
