+++ apps/backend/src/guestAuthTestHarness/index.ts
@@ -1,0 +1,35 @@
+export {
+  DROPPED_ENTITIES_SUPPORTED,
+  DROPPED_ENTITIES_UNSUPPORTED,
+  GUEST_SYNC_NOT_DRAINED,
+  LEGACY_REPLAY_CAPABILITIES,
+  createMergeState,
+  createReviewEventClientEventDedupMergeFixture,
+  createUserSettingsState,
+  createWorkspaceState,
+  hashGuestToken,
+} from "./fixtures";
+export {
+  addWorkspaceMembership,
+  membershipKey,
+  type CardState,
+  type DeckState,
+  type GuestReplicaAliasState,
+  type GuestSessionState,
+  type GuestUpgradeExecutorParam,
+  type GuestUpgradeHistoryState,
+  type HotChangeState,
+  type InstallationState,
+  type MutableState,
+  type ReviewEventClientEventDedupMergeFixture,
+  type ReviewEventState,
+  type UserSettingsState,
+  type WorkspaceMembershipRole,
+  type WorkspaceReplicaState,
+  type WorkspaceState,
+} from "./models";
+export {
+  createGuestUpgradeExecutor,
+  isGuestUpgradeMergeOnlyExecutorQuery,
+} from "./executor";
+export { createQueryResult } from "./queryResult";
