+++ apps/backend/src/guestAuthTestHarness/queryResult.ts
@@ -1,0 +1,11 @@
+import type pg from "pg";
+
+export function createQueryResult<Row extends pg.QueryResultRow>(rows: ReadonlyArray<Row>): pg.QueryResult<Row> {
+  return {
+    command: "SELECT",
+    rowCount: rows.length,
+    oid: 0,
+    fields: [],
+    rows: [...rows],
+  };
+}
